requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.1
django-browser-reload==1.15.0
django-crispy-forms==2.3
django-forms-dynamic==1.0.0
django-formtools==2.3
django-htmx==1.19.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
gunicorn==23.0.0
h11==0.14.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
whitenoise==6.7.0
